App to connect BLE and Analyze Data
Basic/Simple app to connect to the nRF52840 chip. This app also uses whatever data analysis script put in the analyzer.py file to interrupt the data coming in from the microcontroller. Can also graph the data it receives. 

Obviously once we know the specific analysis functions and graphs we want this will be changed. This is just a simple app to test that the microcontroller works with the app and we can pull data and run graphs. 

There are so many files that for now a zip was easiest to upload.